requirements.txt
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dill==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.8
flake8==7.1.1
Flask==3.1.0
h11==0.14.0
httptools==0.6.4
idna==3.10
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.14.1
mypy-extensions==1.0.0
neo4j==5.27.0
oracledb==2.5.1
platformdirs==4.3.6
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
pylint==3.3.4
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.45.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
Werkzeug==3.1.3
requests
chromadb
openai>=1.0.0
fastapi 
uvicorn
gunicorn
